+++ testing/test_full.py
@@ -31,9 +31,17 @@ all_models_segmentation = [
     "msdnet",
 ]
 # pre-defined regression/classification model types for testing
-all_models_regression = ["densenet121", "vgg16", "resnet18", "resnet50"]
+all_models_regression = [
+    "densenet121",
+    "vgg16",
+    "resnet18",
+    "resnet50",
+    "efficientnetb0",
+]
 all_clip_modes = ["norm", "value", "agc"]
 all_norm_type = ["batch", "instance"]
+
+all_model_type = ["torch", "openvino"]
 
 patch_size = {"2D": [128, 128, 1], "3D": [32, 32, 32]}
 
@@ -141,6 +149,12 @@ def test_constructTrainingCSV():
                 i += 1
 
 
+def sanitize_outputDir():
+    if os.path.isdir(outputDir):
+        shutil.rmtree(outputDir)  # overwrite previous results
+    Path(outputDir).mkdir(parents=True, exist_ok=True)
+
+
 def test_train_segmentation_rad_2d(device):
     print("Starting 2D Rad segmentation tests")
     # read and parse csv
@@ -150,25 +164,27 @@ def test_train_segmentation_rad_2d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["class_list"] = [0, 255]
     parameters["model"]["amp"] = True
     parameters["model"]["num_channels"] = 3
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # read and initialize parameters for specific data dimension
     for model in all_models_segmentation:
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
             parameters=parameters,
             device=device,
-            reset_prev=True,
+            resume=False,
+            reset=True,
         )
 
     print("passed")
@@ -190,15 +206,16 @@ def test_train_segmentation_sdnet_rad_2d(device):
     parameters["model"]["class_list"] = [0, 255]
     parameters["model"]["num_channels"] = 1
     parameters["model"]["architecture"] = "sdnet"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
-    shutil.rmtree(outputDir)  # overwrite previous results
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
         parameters=parameters,
         device=device,
-        reset_prev=True,
+        resume=False,
+        reset=True,
     )
 
     print("passed")
@@ -214,36 +231,40 @@ def test_train_segmentation_rad_3d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_3d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["3D"]
     parameters["model"]["dimension"] = 3
     parameters["model"]["class_list"] = [0, 1]
     parameters["model"]["amp"] = True
     parameters["in_memory"] = True
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for model in all_models_segmentation:
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
             parameters=parameters,
             device=device,
-            reset_prev=True,
+            resume=False,
+            reset=True,
         )
 
     print("passed")
 
 
 def test_train_regression_rad_2d(device):
+    print("Starting 2D Rad regression tests")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_regression.yaml", version_check_flag=False
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["amp"] = False
@@ -254,30 +275,33 @@ def test_train_regression_rad_2d(device):
     parameters["model"]["num_channels"] = 3
     parameters["model"]["class_list"] = parameters["headers"]["predictionHeaders"]
     parameters["scaling_factor"] = 1
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for model in all_models_regression:
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
             parameters=parameters,
             device=device,
-            reset_prev=True,
+            resume=False,
+            reset=True,
         )
 
     print("passed")
 
 
 def test_train_brainage_rad_2d(device):
+    print("Starting brain age tests")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_regression.yaml", version_check_flag=False
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["amp"] = False
@@ -289,25 +313,28 @@ def test_train_brainage_rad_2d(device):
     parameters["model"]["class_list"] = parameters["headers"]["predictionHeaders"]
     parameters["scaling_factor"] = 1
     parameters["model"]["architecture"] = "brain_age"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
-    shutil.rmtree(outputDir)  # overwrite previous results
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
         parameters=parameters,
         device=device,
-        reset_prev=True,
+        resume=False,
+        reset=True,
     )
 
     print("passed")
 
 
 def test_train_regression_rad_3d(device):
+    print("Starting 3D Rad regression tests")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_regression.yaml", version_check_flag=False
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["3D"]
     parameters["model"]["dimension"] = 3
     # read and parse csv
@@ -316,26 +343,32 @@ def test_train_regression_rad_3d(device):
     )
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
     parameters["model"]["class_list"] = parameters["headers"]["predictionHeaders"]
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for model in all_models_regression:
+        if "efficientnet" in model:
+            parameters["patch_size"] = [16, 16, 16]
+        else:
+            parameters["patch_size"] = patch_size["3D"]
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
             parameters=parameters,
             device=device,
-            reset_prev=True,
+            resume=False,
+            reset=True,
         )
 
     print("passed")
 
 
 def test_train_classification_rad_2d(device):
+    print("Starting 2D Rad classification tests")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -349,26 +382,28 @@ def test_train_classification_rad_2d(device):
         inputDir + "/train_2d_rad_classification.csv"
     )
     parameters["model"]["num_channels"] = 3
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for model in all_models_regression:
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
             parameters=parameters,
             device=device,
-            reset_prev=True,
+            resume=False,
+            reset=True,
         )
 
     print("passed")
 
 
 def test_train_classification_rad_3d(device):
+    print("Starting 3D Rad classification tests")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -382,25 +417,31 @@ def test_train_classification_rad_3d(device):
     )
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    parameters["model"]["onnx_export"] = False
     # loop through selected models and train for single epoch
     for model in all_models_regression:
+        if "efficientnet" in model:
+            parameters["patch_size"] = [16, 16, 16]
+        else:
+            parameters["patch_size"] = patch_size["3D"]
         parameters["model"]["architecture"] = model
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
             parameters=parameters,
             device=device,
-            reset_prev=True,
+            resume=False,
+            reset=True,
         )
 
     print("passed")
 
 
-def test_inference_classification_rad_3d(device):
+def test_train_resume_inference_classification_rad_3d(device):
+    print("Starting 3D Rad classification tests for resume and reset")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -417,14 +458,43 @@ def test_inference_classification_rad_3d(device):
     # loop through selected models and train for single epoch
     model = all_models_regression[0]
     parameters["model"]["architecture"] = model
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    parameters["model"]["onnx_export"] = False
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
         parameters=parameters,
         device=device,
-        reset_prev=True,
+        resume=False,
+        reset=True,
     )
+
+    ## testing resume with parameter updates
+    parameters["num_epochs"] = 2
+    parameters["nested_training"]["testing"] = -5
+    parameters["nested_training"]["validation"] = -5
+    TrainingManager(
+        dataframe=training_data,
+        outputDir=outputDir,
+        parameters=parameters,
+        device=device,
+        resume=True,
+        reset=False,
+    )
+
+    ## testing resume without parameter updates
+    parameters["num_epochs"] = 1
+    parameters["nested_training"]["testing"] = -5
+    parameters["nested_training"]["validation"] = -5
+    TrainingManager(
+        dataframe=training_data,
+        outputDir=outputDir,
+        parameters=parameters,
+        device=device,
+        resume=False,
+        reset=False,
+    )
+
     parameters["output_dir"] = outputDir  # this is in inference mode
     InferenceManager(
         dataframe=training_data,
@@ -436,7 +506,93 @@ def test_inference_classification_rad_3d(device):
     print("passed")
 
 
+def test_inference_optimize_classification_rad_3d(device):
+    print("Starting 3D Rad segmentation tests for optimization")
+    # read and initialize parameters for specific data dimension
+    parameters = parseConfig(
+        testingDir + "/config_classification.yaml", version_check_flag=False
+    )
+    parameters["modality"] = "rad"
+    parameters["patch_size"] = patch_size["3D"]
+    parameters["model"]["dimension"] = 3
+    # read and parse csv
+    training_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_3d_rad_classification.csv"
+    )
+    parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
+    parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    parameters["model"]["architecture"] = all_models_regression[0]
+    parameters["model"]["onnx_export"] = True
+    sanitize_outputDir()
+    TrainingManager(
+        dataframe=training_data,
+        outputDir=outputDir,
+        parameters=parameters,
+        device=device,
+        resume=False,
+        reset=True,
+    )
+
+    ## testing inference
+    for model_type in all_model_type:
+        parameters["model"]["type"] = model_type
+        parameters["output_dir"] = outputDir  # this is in inference mode
+        InferenceManager(
+            dataframe=training_data,
+            outputDir=outputDir,
+            parameters=parameters,
+            device=device,
+        )
+
+    print("passed")
+
+
+def test_inference_optimize_segmentation_rad_2d(device):
+    print("Starting 2D Rad segmentation tests for optimization")
+    # read and parse csv
+    parameters = parseConfig(
+        testingDir + "/config_segmentation.yaml", version_check_flag=False
+    )
+    training_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_2d_rad_segmentation.csv"
+    )
+    parameters["patch_size"] = patch_size["2D"]
+    parameters["modality"] = "rad"
+    parameters["model"]["dimension"] = 2
+    parameters["model"]["class_list"] = [0, 255]
+    parameters["model"]["amp"] = True
+    parameters["save_output"] = True
+    parameters["model"]["num_channels"] = 3
+    parameters["metrics"] = ["dice"]
+    parameters["model"]["architecture"] = "resunet"
+    parameters["model"]["onnx_export"] = True
+    parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    sanitize_outputDir()
+    TrainingManager(
+        dataframe=training_data,
+        outputDir=outputDir,
+        parameters=parameters,
+        device=device,
+        resume=False,
+        reset=True,
+    )
+
+    ## testing inference
+    for model_type in all_model_type:
+        parameters["model"]["type"] = model_type
+        parameters["output_dir"] = outputDir  # this is in inference mode
+        InferenceManager(
+            dataframe=training_data,
+            outputDir=outputDir,
+            parameters=parameters,
+            device=device,
+        )
+
+    print("passed")
+
+
 def test_inference_classification_with_logits_single_fold_rad_3d(device):
+    print("Starting 3D Rad classification tests for single fold logits inference")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -455,15 +611,15 @@ def test_inference_classification_with_logits_single_fold_rad_3d(device):
     # loop through selected models and train for single epoch
     model = all_models_regression[0]
     parameters["model"]["architecture"] = model
-    if os.path.exists(outputDir):
-        shutil.rmtree(outputDir)  # overwrite previous results
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    parameters["model"]["onnx_export"] = False
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
         parameters=parameters,
         device=device,
-        reset_prev=True,
+        resume=False,
+        reset=True,
     )
     parameters["output_dir"] = outputDir  # this is in inference mode
     InferenceManager(
@@ -477,6 +633,7 @@ def test_inference_classification_with_logits_single_fold_rad_3d(device):
 
 
 def test_inference_classification_with_logits_multiple_folds_rad_3d(device):
+    print("Starting 3D Rad classification tests for multi-fold logits inference")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -487,7 +644,7 @@ def test_inference_classification_with_logits_multiple_folds_rad_3d(device):
     parameters["model"]["final_layer"] = "logits"
     # necessary for n-fold cross-validation inference
     parameters["nested_training"]["validation"] = 2
-
+    parameters["model"]["onnx_export"] = False
     # read and parse csv
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_3d_rad_classification.csv"
@@ -497,13 +654,14 @@ def test_inference_classification_with_logits_multiple_folds_rad_3d(device):
     # loop through selected models and train for single epoch
     model = all_models_regression[0]
     parameters["model"]["architecture"] = model
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
         parameters=parameters,
         device=device,
-        reset_prev=True,
+        resume=False,
+        reset=True,
     )
     parameters["output_dir"] = outputDir  # this is in inference mode
     InferenceManager(
@@ -517,6 +675,7 @@ def test_inference_classification_with_logits_multiple_folds_rad_3d(device):
 
 
 def test_scheduler_classification_rad_2d(device):
+    print("Starting 2D Rad segmentation tests for scheduler")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -532,6 +691,7 @@ def test_scheduler_classification_rad_2d(device):
     parameters["model"]["architecture"] = "densenet121"
     parameters["model"]["norm_type"] = "instance"
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
+    parameters["model"]["onnx_export"] = False
     # loop through selected models and train for single epoch
     for scheduler in global_schedulers_dict:
         parameters["scheduler"] = {}
@@ -546,14 +706,15 @@ def test_scheduler_classification_rad_2d(device):
             outputDir=outputDir,
             parameters=parameters,
             device=device,
-            reset_prev=True,
+            resume=False,
+            reset=True,
         )
 
-    shutil.rmtree(outputDir)
     print("passed")
 
 
 def test_optimizer_classification_rad_2d(device):
+    print("Starting 2D Rad classification tests for optimizer")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -568,6 +729,7 @@ def test_optimizer_classification_rad_2d(device):
     parameters["model"]["num_channels"] = 3
     parameters["model"]["architecture"] = "densenet121"
     parameters["model"]["norm_type"] = "none"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for optimizer in global_optimizer_dict:
@@ -583,14 +745,15 @@ def test_optimizer_classification_rad_2d(device):
             outputDir=outputDir,
             parameters=parameters,
             device=device,
-            reset_prev=True,
+            resume=False,
+            reset=True,
         )
 
-    shutil.rmtree(outputDir)
     print("passed")
 
 
 def test_clip_train_classification_rad_3d(device):
+    print("Starting 3D Rad classification tests for clipping")
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
         testingDir + "/config_classification.yaml", version_check_flag=False
@@ -605,28 +768,28 @@ def test_clip_train_classification_rad_3d(device):
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
     parameters["model"]["architecture"] = "vgg16"
     parameters["model"]["norm_type"] = "None"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for clip_mode in all_clip_modes:
         parameters["clip_mode"] = clip_mode
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        shutil.rmtree(outputDir)  # overwrite previous results
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
             parameters=parameters,
             device=device,
-            reset_prev=True,
+            resume=False,
+            reset=True,
         )
-    shutil.rmtree(outputDir)  # overwrite previous results
     print("passed")
 
 
 def test_normtype_train_segmentation_rad_3d(device):
-    # read and initialize parameters for specific data dimension
     print("Starting 3D Rad segmentation tests for normtype")
+    # read and initialize parameters for specific data dimension
     # read and parse csv
     # read and initialize parameters for specific data dimension
     parameters = parseConfig(
@@ -642,6 +805,7 @@ def test_normtype_train_segmentation_rad_3d(device):
     parameters["save_output"] = True
     parameters["data_postprocessing"] = {"fill_holes"}
     parameters["in_memory"] = True
+    parameters["model"]["onnx_export"] = False
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
@@ -651,15 +815,18 @@ def test_normtype_train_segmentation_rad_3d(device):
             parameters["model"]["norm_type"] = norm
             parameters["nested_training"]["testing"] = -5
             parameters["nested_training"]["validation"] = -5
+            if os.path.isdir(outputDir):
+                shutil.rmtree(outputDir)  # overwrite previous results
             Path(outputDir).mkdir(parents=True, exist_ok=True)
             TrainingManager(
                 dataframe=training_data,
                 outputDir=outputDir,
                 parameters=parameters,
                 device=device,
-                reset_prev=True,
+                resume=False,
+                reset=True,
             )
-            shutil.rmtree(outputDir)  # overwrite previous results
+
         print("passed")
 
 
@@ -672,6 +839,7 @@ def test_metrics_segmentation_rad_2d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["class_list"] = [0, 255]
@@ -680,16 +848,17 @@ def test_metrics_segmentation_rad_2d(device):
     parameters["model"]["num_channels"] = 3
     parameters["metrics"] = ["dice", "hausdorff", "hausdorff95"]
     parameters["model"]["architecture"] = "resunet"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
         parameters=parameters,
         device=device,
-        reset_prev=True,
+        resume=False,
+        reset=True,
     )
-    shutil.rmtree(outputDir)  # overwrite previous results
 
     print("passed")
 
@@ -703,6 +872,7 @@ def test_metrics_regression_rad_2d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_regression.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["class_list"] = [0, 255]
@@ -710,16 +880,17 @@ def test_metrics_regression_rad_2d(device):
     parameters["model"]["amp"] = False
     parameters["model"]["num_channels"] = 3
     parameters["model"]["architecture"] = "vgg11"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
         parameters=parameters,
         device=device,
-        reset_prev=True,
+        resume=False,
+        reset=True,
     )
-    shutil.rmtree(outputDir)  # overwrite previous results
 
     print("passed")
 
@@ -733,6 +904,7 @@ def test_losses_segmentation_rad_2d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["class_list"] = [0, 255]
@@ -741,28 +913,30 @@ def test_losses_segmentation_rad_2d(device):
     parameters["model"]["num_channels"] = 3
     parameters["model"]["architecture"] = "resunet"
     parameters["metrics"] = ["dice"]
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
     for loss_type in ["dc", "dc_log", "dcce", "dcce_logits", "tversky"]:
         parameters["loss_function"] = loss_type
         parameters["nested_training"]["testing"] = -5
         parameters["nested_training"]["validation"] = -5
-        Path(outputDir).mkdir(parents=True, exist_ok=True)
+        sanitize_outputDir()
         TrainingManager(
             dataframe=training_data,
             outputDir=outputDir,
             parameters=parameters,
             device=device,
-            reset_prev=True,
+            resume=False,
+            reset=True,
         )
-        shutil.rmtree(outputDir)  # overwrite previous results
+
     print("passed")
 
 
 def test_config_read():
     print("Starting testing reading configuration")
     # read and parse csv
-    file_config_temp = os.path.join(testingDir, "config_segmentation_temp.yaml")
+    file_config_temp = os.path.join(outputDir, "config_segmentation_temp.yaml")
     # if found in previous run, discard.
     if os.path.exists(file_config_temp):
         os.remove(file_config_temp)
@@ -771,7 +945,7 @@ def test_config_read():
         os.path.abspath(baseConfigDir + "/config_all_options.yaml"),
         version_check_flag=False,
     )
-    parameters["data_preprocessing"]["resize"] = [128, 128]
+    parameters["data_preprocessing"]["resize_image"] = [128, 128]
 
     with open(file_config_temp, "w") as file:
         yaml.dump(parameters, file)
@@ -786,22 +960,82 @@ def test_config_read():
     data_loader = ImagesFromDataFrame(training_data, parameters, True, "unit_test")
     if not data_loader:
         sys.exit(1)
+
+    os.remove(file_config_temp)
+
+    # ensure resize_image is triggered
+    parameters["data_preprocessing"].pop("resample")
+    parameters["data_preprocessing"].pop("resample_min")
+    parameters["data_preprocessing"]["resize_image"] = [128, 128]
+
+    with open(file_config_temp, "w") as file:
+        yaml.dump(parameters, file)
+
+    parameters = parseConfig(file_config_temp, version_check_flag=True)
+
+    training_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_2d_rad_segmentation.csv"
+    )
+    if not parameters:
+        sys.exit(1)
+    data_loader = ImagesFromDataFrame(training_data, parameters, True, "unit_test")
+    if not data_loader:
+        sys.exit(1)
+
+    os.remove(file_config_temp)
+
+    # ensure resize_image is triggered
+    parameters["data_preprocessing"].pop("resize_image")
+    parameters["data_preprocessing"]["resize_patch"] = [64, 64]
+
+    with open(file_config_temp, "w") as file:
+        yaml.dump(parameters, file)
+
+    parameters = parseConfig(file_config_temp, version_check_flag=True)
+
+    training_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_2d_rad_segmentation.csv"
+    )
+    if not parameters:
+        sys.exit(1)
+    data_loader = ImagesFromDataFrame(training_data, parameters, True, "unit_test")
+    if not data_loader:
+        sys.exit(1)
+
+    os.remove(file_config_temp)
+
+    # ensure resize_image is triggered
+    parameters["data_preprocessing"].pop("resize_patch")
+    parameters["data_preprocessing"]["resize"] = [64, 64]
+
+    with open(file_config_temp, "w") as file:
+        yaml.dump(parameters, file)
+
+    parameters = parseConfig(file_config_temp, version_check_flag=True)
+
+    training_data, parameters["headers"] = parseTrainingCSV(
+        inputDir + "/train_2d_rad_segmentation.csv"
+    )
+    if not parameters:
+        sys.exit(1)
+    data_loader = ImagesFromDataFrame(training_data, parameters, True, "unit_test")
+    if not data_loader:
+        sys.exit(1)
+
+    os.remove(file_config_temp)
+
     print("passed")
 
 
 def test_cli_function_preprocess():
     print("Starting testing cli function preprocess")
     file_config = os.path.join(testingDir, "config_segmentation.yaml")
-    file_config_temp = os.path.join(testingDir, "config_segmentation_temp.yaml")
-    # if found in previous run, discard.
-    if os.path.exists(file_config_temp):
-        os.remove(file_config_temp)
-        parameter_pickle_file = os.path.join(outputDir, "parameters.pkl")
-        if os.path.exists(parameter_pickle_file):
-            os.remove(parameter_pickle_file)
+    sanitize_outputDir()
+    file_config_temp = os.path.join(outputDir, "config_segmentation_temp.yaml")
     file_data = os.path.join(inputDir, "train_2d_rad_segmentation.csv")
 
     parameters = parseConfig(file_config)
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["model"]["dimension"] = 2
     parameters["model"]["class_list"] = "[0, 255||125]"
@@ -827,7 +1061,6 @@ def test_cli_function_preprocess():
     # check that the length of training data is what we expect
     assert len(training_data) == 10, "Number of rows in dataframe is not 10"
 
-    shutil.rmtree(outputDir)  # overwrite previous results
     print("passed")
 
 
@@ -836,11 +1069,12 @@ def test_cli_function_mainrun(device):
     parameters = parseConfig(
         testingDir + "/config_segmentation.yaml", version_check_flag=False
     )
-    file_config_temp = os.path.join(testingDir, "config_segmentation_temp.yaml")
+    file_config_temp = os.path.join(outputDir, "config_segmentation_temp.yaml")
     # if found in previous run, discard.
     if os.path.exists(file_config_temp):
         os.remove(file_config_temp)
 
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["num_epochs"] = 1
     parameters["nested_training"]["testing"] = 1
@@ -858,8 +1092,11 @@ def test_cli_function_mainrun(device):
 
     file_data = os.path.join(inputDir, "train_2d_rad_segmentation.csv")
 
-    main_run(file_data, file_config_temp, outputDir, True, device, True)
-    shutil.rmtree(outputDir)  # overwrite previous results
+    main_run(
+        file_data, file_config_temp, outputDir, True, device, resume=False, reset=True
+    )
+    if os.path.isdir(outputDir):
+        shutil.rmtree(outputDir)  # overwrite previous results
     print("passed")
 
 
@@ -890,6 +1127,7 @@ def test_dataloader_construction_train_segmentation_3d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_3d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["3D"]
     parameters["save_training"] = True
     parameters["model"]["dimension"] = 3
@@ -898,17 +1136,19 @@ def test_dataloader_construction_train_segmentation_3d(device):
     parameters["model"]["num_channels"] = len(parameters["headers"]["channelHeaders"])
     parameters["model"]["architecture"] = "unet"
     parameters["weighted_loss"] = False
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
     # loop through selected models and train for single epoch
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
         parameters=parameters,
         device=device,
-        reset_prev=True,
+        resume=False,
+        reset=True,
     )
-    shutil.rmtree(outputDir)  # overwrite previous results
+
     print("passed")
 
 
@@ -1034,6 +1274,7 @@ def test_checkpointing_segmentation_rad_2d(device):
     training_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_rad_segmentation.csv"
     )
+    parameters["modality"] = "rad"
     parameters["patch_size"] = patch_size["2D"]
     parameters["num_epochs"] = 1
     parameters["nested_training"]["testing"] = 1
@@ -1050,14 +1291,16 @@ def test_checkpointing_segmentation_rad_2d(device):
         "hd100_per_label",
     ]
     parameters["model"]["architecture"] = "unet"
+    parameters["model"]["onnx_export"] = False
     parameters = populate_header_in_parameters(parameters, parameters["headers"])
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    sanitize_outputDir()
     TrainingManager(
         dataframe=training_data,
         outputDir=outputDir,
         parameters=parameters,
         device=device,
-        reset_prev=True,
+        resume=False,
+        reset=True,
     )
     parameters["num_epochs"] = 2
     parameters["nested_training"]["validation"] = -2
@@ -1067,9 +1310,9 @@ def test_checkpointing_segmentation_rad_2d(device):
         outputDir=outputDir,
         parameters=parameters,
         device=device,
-        reset_prev=False,
+        resume=False,
+        reset=False,
     )
-    shutil.rmtree(outputDir)  # overwrite previous results
 
     print("passed")
 
@@ -1125,21 +1368,24 @@ def test_one_hot_logic():
     class_list = [0, "1||2||3", 4]
     img_tensor_oh = one_hot(img_tensor, class_list)
     img_tensor_oh_rev_array = reverse_one_hot(img_tensor_oh[0], class_list)
+
+    # check for background
     comparison = (random_array == 0) == (img_tensor_oh_rev_array == 0)
     assert comparison.all(), "Arrays at '0' are not equal"
 
-    random_array_sp = (random_array == 1) + (random_array == 2) + (random_array == 3)
-    img_tensor_oh_rev_array_sp = img_tensor_oh_rev_array == 1
-    img_tensor_oh_rev_array_sp[random_array == 4] = False
-    comparison = random_array_sp == img_tensor_oh_rev_array_sp
-    assert comparison.all(), "Special arrays are not equal"
+    # check last foreground
+    comparison = (random_array == np.max(random_array)) == (
+        img_tensor_oh_rev_array == len(class_list) - 1
+    )
+    assert comparison.all(), "Arrays at final foreground are not equal"
 
-    # check for '4'
-    img_tensor_oh_rev_array_sp = img_tensor_oh_rev_array == 1
-    for i in [1, 2, 3]:
-        img_tensor_oh_rev_array_sp[random_array == i] = False
-    comparison = (random_array == 4) == img_tensor_oh_rev_array_sp
-    assert comparison.all(), "Arrays at '4' are not equal"
+    # check combined foreground
+    combined_array = np.logical_or(
+        np.logical_or((random_array == 1), (random_array == 2)), (random_array == 3)
+    )
+    comparison = combined_array == (img_tensor_oh_rev_array == 1)
+    assert comparison.all(), "Arrays at the combined foreground are not equal"
+
     print("passed")
 
 
@@ -1184,12 +1430,15 @@ def test_anonymizer():
                 shutil.rmtree(file_to_delete)
             else:
                 os.remove(file_to_delete)
+
     print("passed")
 
 
 def test_train_inference_segmentation_histology_2d(device):
     print("Starting histology train/inference tests")
-    output_dir_patches = os.path.join(testingDir, "histo_patches")
+    # overwrite previous results
+    sanitize_outputDir()
+    output_dir_patches = os.path.join(outputDir, "histo_patches")
     if os.path.isdir(output_dir_patches):
         shutil.rmtree(output_dir_patches)
     Path(output_dir_patches).mkdir(parents=True, exist_ok=True)
@@ -1205,6 +1454,7 @@ def test_train_inference_segmentation_histology_2d(device):
     parameters_patch = {}
     # extracting minimal number of patches to ensure that the test does not take too long
     parameters_patch["num_patches"] = 3
+    parameters_patch["patch_size"] = [128, 128]
 
     with open(file_config_temp, "w") as file:
         yaml.dump(parameters_patch, file)
@@ -1232,29 +1482,34 @@ def test_train_inference_segmentation_histology_2d(device):
     parameters["nested_training"]["testing"] = 1
     parameters["nested_training"]["validation"] = -2
     parameters["metrics"] = ["dice"]
-    # overwrite previous results
-    if os.path.isdir(outputDir):
-        shutil.rmtree(outputDir)
-    Path(outputDir).mkdir(parents=True, exist_ok=True)
+    parameters["model"]["onnx_export"] = True
+    modelDir = os.path.join(outputDir, "modelDir")
+    Path(modelDir).mkdir(parents=True, exist_ok=True)
     TrainingManager(
         dataframe=training_data,
-        outputDir=outputDir,
+        outputDir=modelDir,
         parameters=parameters,
         device=device,
-        reset_prev=True,
+        resume=False,
+        reset=True,
     )
-    parameters["output_dir"] = outputDir  # this is in inference mode
     inference_data, parameters["headers"] = parseTrainingCSV(
         inputDir + "/train_2d_histo_segmentation.csv", train=False
     )
 
-    InferenceManager(
-        dataframe=inference_data,
-        outputDir=outputDir,
-        parameters=parameters,
-        device=device,
-    )
-
-    shutil.rmtree(output_dir_patches)
+    for model_type in all_model_type:
+        parameters["nested_training"]["testing"] = 1
+        parameters["nested_training"]["validation"] = -2
+        parameters["output_dir"] = modelDir  # this is in inference mode
+        inference_data, parameters["headers"] = parseTrainingCSV(
+            inputDir + "/train_2d_histo_segmentation.csv", train=False
+        )
+        parameters["model"]["type"] = model_type
+        InferenceManager(
+            dataframe=inference_data,
+            outputDir=modelDir,
+            parameters=parameters,
+            device=device,
+        )
 
     print("passed")